% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_scores}
\alias{calculate_scores}
\title{Calculate scores for plotting}
\usage{
calculate_scores(answers, values)
}
\arguments{
\item{answers}{Answers from a form}

\item{values}{Values tibble}
}
\value{
final_score tibble of scores per value
}
\description{
Calculate scores for plotting
}
