% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_list}
\alias{convert_to_list}
\title{Prepare questions as a list for dynamic generation of form}
\usage{
convert_to_list(questions_tb)
}
\arguments{
\item{questions_tb}{a tibble of questions. Needs id and question columns}
}
\value{
list_form a list for the form
}
\description{
Prepare questions as a list for dynamic generation of form
}
