% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_key}
\alias{create_key}
\title{Combine questions and weights to form a key}
\usage{
create_key(questions, values)
}
\arguments{
\item{questions}{tibble of questions}

\item{weights}{tibble of weights}
}
\value{
key tibble of keys with weights attributed to question ids
}
\description{
Combine questions and weights to form a key
}
