% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_answers}
\alias{load_answers}
\title{Load answers in a long format from a csv}
\usage{
load_answers(csv_path = "inst/tests/tiny_answers.csv")
}
\arguments{
\item{csv_path}{csv path}
}
\value{
answers
}
\description{
Load answers in a long format from a csv
}
