% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_choices}
\alias{load_choices}
\title{Load up choices from a csv}
\usage{
load_choices(csv_path = "inst/PVQ_choices.csv", lang = "fr")
}
\arguments{
\item{csv_path}{csv path}

\item{lang}{pick which language}
}
\value{
choices tibble of language filtered choices
}
\description{
Load up choices from a csv
}
