% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_values}
\alias{load_values}
\title{Load up values from a csv}
\usage{
load_values(csv_path = "inst/PVQ_values.csv", pvq_type = 21)
}
\arguments{
\item{csv_path}{csv path}

\item{pvq_type}{either 21 or 40}
}
\value{
values tibble of filtered values
}
\description{
Load up values from a csv
}
