% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_radar}
\alias{plot_radar}
\title{Generate a radar plot from data}
\usage{
plot_radar(scores)
}
\arguments{
\item{scores}{a named vector of results}
}
\value{
ggplot
}
\description{
Generate a radar plot from data
}
